refactor: move pow method to exponential file
